// src/schemas/general-configuration-schema/types.ts
export enum ConfigurationType {
  COMMISSION = "commission",
  VALIDATION_RULE = "validation_rule",
  ENABLE_DISABLE = "enable_disable",
}

export enum CommissionType {
  PERCENTAGE = "Percentage",
  LUMP_SUM = "LumpSum",
}

export enum ConfigurationKey {
  FIB_COMMISSION = "FIB_COMMISSION",
  FASTPAY_COMMISSION = "FASTPAY_COMMISSION",
  CREDIT_COMMISSION = "CREDIT_COMMISSION",
  MAX_OUTSTANDING_INVOICES_ELECTRICITY = "MAX_OUTSTANDING_INVOICES_ELECTRICITY",
}

export enum ServiceType {
  ELECTRICITY = "electricity",
}